Added shortest path algorithm & misc utility methods
Added an implementation of an A* shortest path algorithm. 
Added an 'drain' method that unfloods the maze each turn
Added a getMessage method
Found that I had to move the floodfill call to flood in all circumstances, rather than only when the target was not clear. I'm not sure why this is necessary and didn't have time to investigate. 
The solution works for 3 simple mazes testes but fails on the first complex maze (eventually finds the current location as a target (furthest known node ?) and hence fails to generate a path, or a next move)