Added explore ability if time available
Added an explore condition to the findTarget() method , which activates if the control room is found but there is still ample time available. 
This helps with mapping completeness and finding the true shortestPath home. 
Solution now passes tests 1-6, but fails test 7 & 8. In both cases, the solution takes too long to find the control room. A great deal of time is wasted traversing from one "furthest point" to the next " furthest point" rather than exploring rooms.